ml-model/requirements.txt
pandas==1.*
numpy==1.*
scikit-learn==1.*
joblib==1.*
google-cloud-bigquery==3.*
google-cloud-aiplatform==1.* 